Revert "[APP-505] Optionally use gRPC service to report logs (#1198)"
This reverts commit 1d0c07644c35708df57a7e82af35d5ff127e317c.
