+++ web/server/entrypoint.go
@@ -2,25 +2,254 @@
 package server
 
 import (
+	"bytes"
 	"context"
+	"encoding/json"
+	"fmt"
 	"net"
+	"net/http"
 	"os"
 	"runtime/pprof"
+	"sync"
 	"time"
 
 	"github.com/edaniels/golog"
 	"github.com/edaniels/gostream/codec/x264"
+	"github.com/pkg/errors"
 	"go.opencensus.io/trace"
 	"go.uber.org/multierr"
+	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
 	"go.viam.com/utils"
 	"go.viam.com/utils/perf"
 	"go.viam.com/utils/rpc"
 
 	"go.viam.com/rdk/config"
+	"go.viam.com/rdk/rlog"
 	robotimpl "go.viam.com/rdk/robot/impl"
 	"go.viam.com/rdk/robot/web"
 	weboptions "go.viam.com/rdk/robot/web/options"
 )
+
+type wrappedLogger struct {
+	base  zapcore.Core
+	extra []zapcore.Field
+}
+
+func (l *wrappedLogger) Enabled(level zapcore.Level) bool {
+	return l.base.Enabled(level)
+}
+
+func (l *wrappedLogger) With(f []zapcore.Field) zapcore.Core {
+	return &wrappedLogger{l, f}
+}
+
+func (l *wrappedLogger) Check(e zapcore.Entry, ce *zapcore.CheckedEntry) *zapcore.CheckedEntry {
+	return l.base.Check(e, ce)
+}
+
+func (l *wrappedLogger) Write(e zapcore.Entry, f []zapcore.Field) error {
+	field := []zapcore.Field{}
+	field = append(field, l.extra...)
+	field = append(field, f...)
+	return l.base.Write(e, field)
+}
+
+func (l *wrappedLogger) Sync() error {
+	return l.base.Sync()
+}
+
+func newNetLogger(config *config.Cloud) (*netLogger, error) {
+	hostname, err := os.Hostname()
+	if err != nil {
+		return nil, err
+	}
+
+	cancelCtx, cancel := context.WithCancel(context.Background())
+	nl := &netLogger{
+		hostname:  hostname,
+		cfg:       config,
+		cancelCtx: cancelCtx,
+		cancel:    cancel,
+	}
+	nl.activeBackgroundWorkers.Add(1)
+	utils.ManagedGo(nl.backgroundWorker, nl.activeBackgroundWorkers.Done)
+	return nl, nil
+}
+
+type netLogger struct {
+	hostname   string
+	cfg        *config.Cloud
+	httpClient http.Client
+
+	toLogMutex sync.Mutex
+	toLog      []interface{}
+
+	cancelCtx               context.Context
+	cancel                  func()
+	activeBackgroundWorkers sync.WaitGroup
+}
+
+func (nl *netLogger) queueSize() int {
+	nl.toLogMutex.Lock()
+	defer nl.toLogMutex.Unlock()
+	return len(nl.toLog)
+}
+
+func (nl *netLogger) Close() {
+	// try for up to 10 seconds for log queue to clear before cancelling it
+	for i := 0; i < 1000; i++ {
+		if nl.queueSize() == 0 {
+			break
+		}
+
+		time.Sleep(10 * time.Millisecond)
+	}
+	nl.cancel()
+	nl.activeBackgroundWorkers.Wait()
+	nl.httpClient.CloseIdleConnections()
+}
+
+func (nl *netLogger) Enabled(zapcore.Level) bool {
+	return true
+}
+
+func (nl *netLogger) With(f []zapcore.Field) zapcore.Core {
+	return &wrappedLogger{nl, f}
+}
+
+func (nl *netLogger) Check(e zapcore.Entry, ce *zapcore.CheckedEntry) *zapcore.CheckedEntry {
+	return ce.AddCore(e, nl)
+}
+
+func (nl *netLogger) Write(e zapcore.Entry, f []zapcore.Field) error {
+	// TODO(erh): should we put a _id uuid on here so we don't log twice?
+
+	for idx, ff := range f {
+		if ff.String == "" && ff.Interface != nil {
+			ff.String = fmt.Sprintf("%v", ff.Interface)
+			f[idx] = ff
+		}
+	}
+
+	x := map[string]interface{}{
+		"id":     nl.cfg.ID,
+		"host":   nl.hostname,
+		"log":    e,
+		"fields": f,
+	}
+
+	nl.addToQueue(x)
+
+	if e.Level == zapcore.FatalLevel || e.Level == zapcore.DPanicLevel || e.Level == zapcore.PanicLevel {
+		// program is going to go away, let's try and sync all our messages before then
+		return nl.Sync()
+	}
+
+	return nil
+}
+
+func (nl *netLogger) addToQueue(x interface{}) {
+	nl.toLogMutex.Lock()
+	defer nl.toLogMutex.Unlock()
+
+	if len(nl.toLog) > 20000 {
+		// TODO(erh): sample?
+		nl.toLog = nl.toLog[1:]
+	}
+	nl.toLog = append(nl.toLog, x)
+}
+
+func (nl *netLogger) writeToServer(x interface{}) error {
+	js, err := json.Marshal(x)
+	if err != nil {
+		return err
+	}
+
+	// we specifically don't use a parented cancellable context here so we can make sure we finish writing but
+	// we will only give it up to 5 seconds to do so in case we are trying to shutdown.
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second*5)
+	defer cancel()
+	r, err := http.NewRequestWithContext(ctx, http.MethodPost, nl.cfg.LogPath, bytes.NewReader(js))
+	if err != nil {
+		return errors.Wrap(err, "error creating log request")
+	}
+	r.Header.Set("Secret", nl.cfg.Secret)
+
+	resp, err := nl.httpClient.Do(r)
+	if err != nil {
+		return err
+	}
+	defer func() {
+		utils.UncheckedError(resp.Body.Close())
+	}()
+	return nil
+}
+
+func (nl *netLogger) backgroundWorker() {
+	normalInterval := 100 * time.Millisecond
+	abnormalInterval := 5 * time.Second
+	interval := normalInterval
+	for {
+		cancelled := false
+		if !utils.SelectContextOrWait(nl.cancelCtx, interval) {
+			cancelled = true
+		}
+		err := nl.Sync()
+		if err != nil && !errors.Is(err, context.Canceled) {
+			interval = abnormalInterval
+			// fall back to regular logging
+			rlog.Logger.Infof("error logging to network: %s", err)
+		} else {
+			interval = normalInterval
+		}
+		if cancelled {
+			return
+		}
+	}
+}
+
+func (nl *netLogger) Sync() error {
+	// TODO(erh): batch writes
+
+	for {
+		x := func() interface{} {
+			nl.toLogMutex.Lock()
+			defer nl.toLogMutex.Unlock()
+
+			if len(nl.toLog) == 0 {
+				return nil
+			}
+
+			x := nl.toLog[0]
+			nl.toLog = nl.toLog[1:]
+
+			return x
+		}()
+
+		if x == nil {
+			return nil
+		}
+
+		err := nl.writeToServer(x)
+		if err != nil {
+			nl.addToQueue(x) // we'll try again later
+			return err
+		}
+	}
+}
+
+func addCloudLogger(logger golog.Logger, cfg *config.Cloud) (golog.Logger, func(), error) {
+	nl, err := newNetLogger(cfg)
+	if err != nil {
+		return nil, nil, err
+	}
+	l := logger.Desugar()
+	l = l.WithOptions(zap.WrapCore(func(c zapcore.Core) zapcore.Core {
+		return zapcore.NewTee(c, nl)
+	}))
+	return l.Sugar(), nl.Close, nil
+}
 
 // Arguments for the command.
 type Arguments struct {
@@ -28,6 +257,7 @@ type Arguments struct {
 	ConfigFile         string `flag:"0,required,usage=robot config file"`
 	CPUProfile         string `flag:"cpuprofile,usage=write cpu profile to file"`
 	Debug              bool   `flag:"debug"`
+	LogURL             string `flag:"logurl,usage=url to log messages to"`
 	SharedDir          string `flag:"shareddir,usage=web resource directory"`
 	Version            bool   `flag:"version,usage=print version"`
 	WebProfile         bool   `flag:"webprofile,usage=include profiler in http server"`
@@ -73,7 +303,7 @@ func RunServer(ctx context.Context, args []string, logger golog.Logger) (err err
 	}
 	cancel()
 
-	if cfg.Cloud != nil && (cfg.Cloud.LogPath != "" || cfg.Cloud.AppAddress != "") {
+	if cfg.Cloud != nil && cfg.Cloud.LogPath != "" {
 		var closer func()
 		logger, closer, err = addCloudLogger(logger, cfg.Cloud)
 		if err != nil {
